[svn r13] - Fixed bug ray_n on polygon.c and obstacle_avoidance.c. - Back to prescale value 1 in pwm_mc.c and chnged define  MOD3_XX to CH3_XXX.

+++ modules/devices/robot/obstacle_avoidance/obstacle_avoidance.h
@@ -64,7 +64,7 @@ struct obstacle_avoidance {
 
 
 	
-	uint8_t ray_n;
+	uint16_t ray_n;
 	uint8_t cur_poly_idx;
 	uint8_t cur_pt_idx;
